moved code for unit conversion to device base. added unit conversion to messaging device.
